% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleData.R
\docType{package}
\name{FlexibleChoraleHamonicAnalysisPackage}
\alias{FlexibleChoraleHarmonicAnalysis}
\alias{FlexibleChoraleHarmonicAnalysis-package}
\title{Flexible Chorale Harmonic Analysis}
\description{
This package is a "flexible" dataset of harmonic analyses of baroque chorale music by J.S. Bach and M. Praetorius.
What makes it "flexible" is that it doesn't just include ONE analysis of the chorales, but a system for generating
countless valid analyses of the chorales, based on constraints you choose.
}
\details{
This package is a companion to the paper presented at the International
Society of Music Information Retrieval (ISMIR) 2018 conference entitled "A Flexible Approach to Automated
Harmonic Analysis: Multiple Annotations of Chorales by Back and Praetorius" (Condit-Schultz, Ju, and Fujinaga).
See the paper \url{http://ismir2018.ircam.fr/doc/pdfs/283_Paper.pdf} for more details!
}
\section{Raw Data}{


The \code{FlexibleChoralHarmonicAnalysis} package includes lots of useful data.
The original 572 "raw" humdrum/kern files, representing 571 chorales, are included in the package: 371 chorales by Bach and 200 by Praetorius.
(Praetorius' 130th chorale is divided between two files, resulting in one additional file compared to the number of chorales).
You can inspect individual files using the \code{\link{viewHumdrum}} command.
The directory where the raw data files are located on your computer is contained in the \code{\link{humdrumDirectory}} variable.
The actual paths to all 572 files are stored in the \code{\link{humdrumPaths}} variable.

In addition to the "raw" data, the package contains various features that have already been parsed from the humdrum data.
The parsed data comes in two equivalent forms, "by slice" and "by note", stored in the \code{\link{ChoraleTable_Slices}} and \code{\link{ChoraleTable_Notes}} data.table objects respectively.
}

\section{Generating Harmonic Analyses}{


To generate specific harmonic (roman numeral) analyses of the 572 chorales, use the function \code{\link{analyzeChorales}}.
Once you've generated analyses, you can view them with the \code{\link{viewHumdrum}} command, or write them to new
humdrum files using the \code{\link{writeHumdrum}} command.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natsguitar/FlexibleChoraleHarmonicAnnotations}
}

}
\author{
\strong{Maintainer}: Nathaniel Condit-Schultz \email{natsguitar@gmail.com}

}
\keyword{internal}
