% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\docType{data}
\name{humdrumDirectory}
\alias{humdrumDirectory}
\title{Raw humdrum data directory}
\format{
An object of class \code{character} of length 1.
}
\usage{
humdrumDirectory
}
\description{
This character string stores the directory where the package's raw humdrum data is stored on your machine.
}
\keyword{datasets}
