% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\docType{data}
\name{humdrumPaths}
\alias{humdrumPaths}
\title{Raw humdrum data paths.}
\format{
An object of class \code{character} of length 572.
}
\usage{
humdrumPaths
}
\description{
This character vector stores the path to all 572 of the package's raw humdrum data files.
}
\keyword{datasets}
