% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleAnalysis.R
\name{\%ranks\%}
\alias{\%ranks\%}
\title{Apply arbitrary ranks to values}
\usage{
x \%ranks\% patterns
}
\arguments{
\item{x}{A vector. Usually the \code{Chord$Qualities} field. This field will always be coerced to a character vector.}

\item{patterns}{A named vector of numbers.}
}
\description{
The \verb{\%ranks\%} infix function allows you to apply arbitrary ranks to an input vector.
This is mainly used with the \code{Chord$Qualities} field.
}
\details{
Each of the names of the right-side \code{patterns} argument is treated as one or more character strings to match exactly the left-side input vector \code{x}.
The output is assigned the value of the last \code{pattern} that matches.
Multiple patterns can be given the same weight by putting them in the same name, but separated by \code{"|"}.

If one, unnamed, pattern value is included on the right-side, it is used as the default value for anything that matches no patterns.
}
\examples{

test <- c('M', 'm', 'd', 'Mm', 'mm', 'MM', 'P5')

test \%ranks\% c('M' = 5, 'm|d' = 4, 'Mm' = 3, 'mm|MM' = 2, 0)

# returns: c(5, 4, 4, 3, 2, 2, 0)


analyzeChorales(Chord ~ min(Qualities \%ranks\% c('M' = 5, 'm|d' = 4, 'Mm' = 3, 'mm|MM' = 2, 0)))

}
