% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\name{viewHumdrum}
\alias{viewHumdrum}
\title{Inspect a humdrum file from the dataset.}
\usage{
viewHumdrum(composer = "Bach", chorale = 1, analyses = NULL)
}
\arguments{
\item{composer}{Character string: The name of the desired composer (Bach or Praetorius). Partial matches are allowed; Case insensitive.}

\item{chorale}{Numeric (integer): The desired chorale number(s).}

\item{analyses}{An \code{analyses} object, created by \code{\link{analyzeChorales}}. If \code{NULL} the original humdrum files are printed.
\link{analyzeChorales} outputs a single analysis, but multiple analyses can be concatenated and fed to \code{viewHumdrum}:
i.e., \code{viewHumdrum('bach', 10, c(analysis1, analysis2))}}
}
\value{
A list of character vector: one for each viewed chorale. Each element is a vector of humdrum
records/slices.
}
\description{
This function prints humdrum data in the R console.
It can be used to inspect the "raw" humdrum files from the dataset, or to view the annotated humdrum generated by the package.
When this function is called, the humdrum file(s) are printed on the screen.
}
\examples{
viewHumdrum('Bach', 11)
# Prints the 11th Bach chorale, unanalyzed.

viewHumdrum('Praetorius', 1:10)
# Prints the first 10 Praetorius chorales, unanalyzed.

viewHumdrum('Pr', 1:10, analyses = analyzeChorales(rank = list(Chord ~ Count)))
# Prints the first 10 Praetorius chorales, with a particular harmonic analysis.


}
