% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChoraleOutput.R
\name{writeHumdrum}
\alias{writeHumdrum}
\title{Write annotated humdrum data to files}
\usage{
writeHumdrum(analyses, path = NULL, label = NULL)
}
\arguments{
\item{analyses}{An \code{analyses} object, created by \code{\link{analyzeChorales}}.
\link{analyzeChorales} outputs a single analysis, but multiple analyses can be concatenated and fed to \code{writeHumdrum}:
i.e., \code{writeHumdrum(c(analysis1, analysis2))}}

\item{path}{A single character string indicating the file path to write the files into.
Defaults to \code{NULL}, in which case files are written to the current working directory.}

\item{label}{A single character string to append to each filename.
Defaults to \code{NULL}, which causes no label to be appended. Use this to differentiate different batches
of analyses you generate.}
}
\value{
A \code{data.table} with two columns (\code{Records} and \code{FileName}) and \code{nrow(ChoraleTables_Slices)} rows.
\code{Records} contains a character vector of new humdrum records.
\code{FileName} a character vector indicating the new humdrum filenames, with the added \code{path} and \code{label},
for each record/slice.
}
\description{
This function writes new humdrum data (original **kern with added harmonic
analyses spines) to text files.
The original krn file names are used for the output files with the addition an optional label specified by the user and a different extension (\code{.hum}).
}
